Adds documentation index.
The documentation index is a table of document names and their
respective first paragraphs.  This should help a bit in document
navigation, especially as the number of documents grows.

Also added a navigation link from README.md, as this is the likely
project entry point for at least a little while.

Fixes: #56