Implements ecma402_traits with icu4x.
This allows (almost) seamless switching between `rust_icu` and
icu4x for plural formatting.

The API surface is a bit different with ecma402 than icu4x, which
may be a necessary concession to API uniformity.

For the time being, the significant digits in the formatted number
are not handled completely, so, for example, excessive precision will
not be handled correctly.

Also for the time being, no other APIs defined in `ecma402_traits` have
a icu4x implementation.  Plural rules formatting was chosen as an easy
demo of the common ECMA402 API.