Implements From<FixedDecimal> for PluralOperands (#278)
* Implements From<FixedDecimal> for PluralOperands

This will allow us to convert FixedDecimal representation without loss
of precision into PluralOperands, for correct plural selection.

For example, a number `25` may sometimes be pluralized differently from
`25.0`.

Added the benchmarks for the conversion, though without a baseline it is
not just as useful yet.  (Until we try to reimplement.)

See issue #190.

* Rewrite eq() to branchless

* Adds a test for eq()

We needed to add a custom implementation for eq() to account
for loss of precision in PluralOperands.

* fixup:first

* Clean up the code for From

* Adds more illustrative naming in From

* Makes clippy happy

* Pulls num_fractional_digits out of the loop

It is possible to compute it based on the low end of the magnitude
range.  No performance change per benchmark.

* fixup: benchmark was wrong

* fixup: moves new operand tests to json

Defines new tests for the data model for conversion, and places them
into the JSON files instead of inline with the tests.

* fixup: moves the benchmark loop in

This allows criterion to run the benchmark loop
with a specific time limit.

* fixup: formatting

* fixup: adds individual sample measurements

Helps smoke out specific performance regressions.