Clarify the terms "bag" and "options" for the DateTimeFormat docs
I found myself confused by these terms, and how they worked. I updated
the docs to try and clarify what these represent, and to clarify the
state of the future planned work. It took me a bit of time to figure out
that Components wasn't actually implemented yet, and so this is an
attempt to add more clarity.