Add a tool to inject dhat instrumentation into a file
This is a partial commit of the new tool.

The following commit completes the tool by running an example with the
injected code. dhat-rs must be manually instrumented in rust, as opposed
to other memory tools. The benefit with using this approach is that it
works on macOS, Windows, and Linux, while Valgrind only works on Linux.

The added code takes the strategy of parsing the Rust AST, and injecting
the proper AST for the new code. The rust analyzer project has a pattern
matching search and replace, which was a good candidate, but it wasn't
well documented, and not great for matching the first expression in the
main function.