Optimize likely subtags
This breaks up the current monolithic likely subtags data into four tables
based upon the subtags which are used for searching by the maximize algorithm.
Because of this change, searching can be done using references to the subtags
of the input Locale rather than creating a copy of it.

The result of a search always shares data with the search key. This change
takes advantage of this fact to only store the delta between the search key
and the result.

The performance impact is as follows:
```
likelysubtags/maximize  time:   [1.5220 us 1.5240 us 1.5262 us]
                        change: [-41.398% -41.032% -40.721%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Memory use, as calculated by dhat, is also improved. With these changes:
```
dhat: Total:     221,715 bytes in 88 blocks
dhat: At t-gmax: 108,539 bytes in 30 blocks
dhat: At t-end:  0 bytes in 0 blocks
```

And without the changes:
```
dhat: Total:     286,547 bytes in 67 blocks
dhat: At t-gmax: 140,795 bytes in 27 blocks
dhat: At t-end:  0 bytes in 0 blocks
```

Which shows approximate a 60k improvement in total memory use when running an
example based upon the existing likely subtags benchmark.

Fixes #468.