Add memory benchmarks (#446)
* Add dhat as a dev dependency to instrumented components

* Update .gitignore for the newly generated files

/benchmarks store all of the benchmark data.

* Add a tool to inject dhat instrumentation into a file

This is a partial commit of the new tool.

The following commit completes the tool by running an example with the
injected code. dhat-rs must be manually instrumented in rust, as opposed
to other memory tools. The benefit with using this approach is that it
works on macOS, Windows, and Linux, while Valgrind only works on Linux.

The added code takes the strategy of parsing the Rust AST, and injecting
the proper AST for the new code. The rust analyzer project has a pattern
matching search and replace, which was a good candidate, but it wasn't
well documented, and not great for matching the first expression in the
main function.

* Finish memory_bench so that it automatically runs examples

This commit completes the memory_bench tool. It can be run locally, but
is intended to run in CI. The next commit will add the CI functionality.

* Add a CI job to collect memory benchmarks

This uses a custom fork of the benchmarking tool to add support for
ndjson as a tool. This ndjson uses the same format as the internal json
structure of the benchmarking tool. This allows for fully customizing
the output of the data, and what information is collected.

* Create tools directory

* Add macro support for the memory instrumentation

* Remove the code injection, and update the benchmark tool

* Update the workflows

* Update the Cargo.lock

* Make the DateTime bags more explicit (#429)

* Make the DateTime bags more explicit

I found myself confused on the usage of the bags and the difference
between them and the options provided to the date time. I felt that it
was better to be explicit in the examples, rather than giving more
terse, but potentially misleading examples.

* Address feedback on adding into() examples

* Remove use icu_benchmark_macros statements

* Add mention of the firefox profiler