Optimize likely subtags (#469)
This breaks up the current monolithic likely subtags data into multiple tables
based upon the subtags which are used for searching by the maximize algorithm.
Because of this change, searching can be done using references to the subtags
of the input Locale rather than creating a copy of it.

The result of a search always shares data with the search key. This change
takes advantage of this fact to only store the delta between the search key
and the result.