Support creating a FixedDecimal from a float value through Ryu
This introduces the `FixedDecimal::from_float_ryu` API, which uses Ryu to convert a floating value (`f32` or `f64`, as defined by the impls of `ryu::Float`) into a `FixedDecimal` by first converting the value into a string representation.

To avoid introducing unnecessary dependencies, this API and its dependency on Ryu are hidden behind the `ryu_decimal` feature.

Fixes #166

TODO:
- [ ] Figure out if scientific notation is handled correctly