Add subtag accessors to Locale C FFI
This adds subtag accessors to the Locale FFI to support the accessors
specified on ecma402 Intl.Locale objects.

I think the FFI is the right place for these to live: internal ICU4x callers
can access these values directly, and external callers would otherwise be
forced to reparse the string representation of the locale which seems
less than optimal.

For the unicode extensions, it is less clear whether we want to have
individual accessor methods for the extensions used in ecma402 or a
generic accessor for extensions. It is likely slight more performant to
have individual accessors.

This also adds a return type enum to make it possible to distinguish
between undefined subtags and extensions, and errors while writing.

Fixes #757.