Simplify the parsing logic of LineBreak.txt
Unified the logic setting line breaking class. Before this patch,
"CP_EA" tweak was only set when we parsed a range of codepoints, but we
really should also apply it when we parsed a single codepoint, although
currently the path doesn't do anything with current EastAsianWidth.txt
data.

This is a refactor and shouldn't change the behavior.