Clean up type fallback in TimeZonesProvider
The old way of handling type fallback in the DataProvider made assumptions
about the shape of the CLDR data (that I realize now were wrong, in at least one case).

This adds an explicit function to handle type fallback exactly as specified by UTS-35,
improving the readability of the code.

Updates the data and test case to fix the prior mistake.