Add fr-BE as a test locale.
The latter uses non default weekdata which is useful for week-of-year/month
tests.

I've looked at replacing fr with fr-BE to keep the number of locales
constant but fails due to plurals data not having locale regions. This should
be fixed by #834 which will also IIUC add all regional variants to testdata
anyways.