Replace break_iterator_impl macro with a helper trait (#1548)
The goal of this patch is to remove the macro and preserve the APIs as much
as possible. We'll start simplify things in the later patches.

The removed get_break_property_{utf32,latin1,utf8} helper functions all convert
the character into 32-bit unsigned integer to query the property table. Since
RuleBreakType::CharType is bounded by Into<u32>, we can implement a generic
get_break_property() directly in RuleBreakIterator.

This patch shouldn't change the behavior.