Add icu_segmenter_provider crate
The purpose of the crate is to deserialize the segmenter rule break TOML files
into `SegmenterRuleTable`, and transform it into `RuleBreakDataV1`. The main
function where the transformation takes place is in
SegmenterRuleProvider::generate_break_data(), which is ported from
`generate_rule_segmenter_table` along with many other helpers in
`build.rs`.

Flatten `RuleBreakPropertyTable` into a linear structure so that it can be
serialize/dezerialize via ZeroVec.

In the next commit, we'll convert line segmenter to use RuleBreakDataV1. This
patch removes "provider_serde" cfg for LineBreakDataV1 just to build
successfully.