Uses tries instead of inversion lists for normalization data (#2235)
* Remove inversion list usage from decomposing normalization

The "decomposition starts with a non-starter" property is stored as
a special marker value in the NFD trie.

* Remove inversion list usage from the composing normalizer

A trie performs better than a fragmented inversion list.

If an app wants to optimize data size at the expense of performance,
it's valid to use a passthrough set that passes more through. For
example, it's valid to use nfkc.toml for both NFC and NFKC.
It's also valid to use passthroughnop.toml (provided but not used
by this changeset) in place of any of nfc.toml, nfkc.toml, or
uts46.toml.

Furthermore, the sets are provided separately instead of being combined
into one trie, because a future change might be able to do away with
nfkc.toml and uts46.toml by using the information available in nfc.toml
and nfkd.toml/uts46d.toml.

The field `first` in the data is expected to be used by a future
change.