Swap the logical halves of the the 32-bit normalization trie values
This way, special marker values become close to zero (most notably
an upcoming marker value can be assigned to 1) when the 32-bit value
is compared as a whole.

This change leaves some variables badly named in order to make the
diff show only the swapping of the halves without any other clean-up.