Switch datagen over to prettyplease
Datagen can't rely on rustfmt existing in the PATH.

On the plus side, `prettyplease` has no qualms about  formatting long lines (rustfmt bails), so some of our databake has gotten prettier. Also according to benchmarks on `prettyplease` it's faster (though it's not apples-to-apples)

On the minus side, `prettyplease` doesn't group slice  entries on the same line for large slices, so it can be a bit vertically sparse. I don't think this is too important.

Ideally we can get a patch release with this so I can use it in Google3, however we may be able to just let this ride the trains until later.