Add some Debug implementations in datagen
This works around the `missing_debug_implementations` regression in https://github.com/rust-lang/rust/issues/111359 . I don't know if that will be fixed soon; I spent some time investigating it and have an idea of what's going on but not enough to fix it (and don't want to spend more time on this right now). So for now, let's just listen to the lint and add these impls.