Remove `syn` from databake
Non-proc-macro deps are compiled separately; and this dep brings `syn` out of the shadow realm of proc-macro deps into the real world, linking it into the final artifact (though it goes almost entirely unused). As it is slow to compile this should get us a slight compilation time win when building in datagen mode.