Using `typedoc` instead of Diplomat-generated docs (#4371)
The JS Diplomat docs are quite bad, they use a module structure which is
not present in JS, and don't have a lot of detail. Generating docs from
the actual code is much more accurate, looks nicer, has more detail, and
doesn't require us to maintain the Diplomat docs generator.