Add YearInfo to CalendarArithmetic (but don't use it yet in any calendar) (#4407)
This is mostly a scaffolding PR implementing the first few steps of
https://github.com/unicode-org/icu4x/issues/3933#issuecomment-1830962051.

This:

- Adds a YearInfo type that is stored in CalendarArithmetic and needs to
be computed whenever the type is created or has its offset updated
- Requires `.offset()` to be called with an `&impl
PrecomputedDataSource<YearInfo>`. An intermediate state of the PR had a
separate PrecomputedDataSource associated type as per the original
design, but I realized it wasn't necessary and it had a tendency to lead
to worse trait resolution issues with the ChineseBased blanket impl
- For now, all construction methods for CalendarArithmetic require
YearInfo to be `()`. A PR actually using this can add the APIs it needs
- As a cleanup I removed `new_with_lunar_ordinals()`. We weren't
maintaining that distinction correctly; and I considered fixing it, but
that method only exists in the case of future needs and when those needs
come we can easily re-add it and do the proper refactoring.

Can be reviewed commit-by-commit if desired, though I'd recommend
reviewing the whole PR at once. It's probably sufficient to review
calendar_arithmetic.rs and chinese_based.rs.