Add explicit lifetime to `ScriptExtensionsSet` impl (#4603)
By eliding this lifetime, the compiler assumes that the iterator is
valid for the lifetime of self, but it it really valid for the lifetime
of the underlying data. An explicit lifetime makes this clear, and
allows for the returned iterator to outlive the `&self` reference, which
is useful in many cases.

I've only fixed this once instance (and not looked for others) since
this is the case that I ran into first; I assume that this pattern is
repeated elsewhere, and it should probably be fixed systematically,
unless there is a rationale for this limitation that I am overlooking.

In any case I'm making the minimal PR here, as it's my first to the
project. Happy to make any changes or additions as requested.

Thanks!