Add a lossy constructor for tinystr (#4691)
An instance where it is easier for me to write the patch and open a PR
than to start a discussion issue.

I want to be able to write,

```rust
const COMMON_SCRIPT: TinyStr<4> = TinyStr::some_constructor(b"Zyyy");
```

but as far as I can tell, the only way to do this currently would
require the unsafe `from_bytes_unchecked` constructor. I would like to
be able to avoid unsafe, as it introduces additional code review
obligations.


I feel like one possible solution to this would be a lossy constructor,
that handles invalid input by substituting invalid bytes with a valid
replacement.


This PR includes one possible version of this. I'm not at all married to
this particular implementation; it is offered as an illustration.