Remove CI download cache (#4871)
It's not clear that downloading from the cache is significantly faster
than downloading from the sources, which are also located on GH's
servers. The download cache however means that we have to worry about it
getting stale (happened with ICU 75), and also the download path is not
exercised in main or nightly CI, because the cache always hits.