Webpack fetch URL fix for Windows
It's probably a \ vs / problem.

new URL('dist', import.meta.url).pathname doesn't work for Windows. After futzing around with different approaches, I found the simplest was just removing file:/// from the full href.

I'd be curious to know if this works on other systems, and I'm open to workshopping a cross-platform solution if not!