Unifying `DataLocale` construction (#4912)
In preparation for #3632 

There are two cases:
* `DataLocales` that are passed into ICU4X constructors:
* These are constructed with `locale!("foo").into()`, or `"foo".parse()`
* When changing a constructor to preferences, the code doesn't need to
be changed if we add a `From<Locale>` impl for the preferences
* `DataLocales` that are put into the `DataRequest::locale` field.
* These are constructed with `langid!("foo").into()`, or `"foo".parse()`
* When changing the field to `&LanguageIdentifier`, these can be
clippy-cleaned-up, as the `.into()` will become redundant.

I've taken care to avoid having intermediate `Locale` or
`LanguageIdentifier` variables, everything that is macro-constructed is
immediately `into()`ed. This will simplify find-and-replace later.