Move `DatagenProvider` into its own crate (#5016)
#4721 

`provider/datagen/src/{provider.rs, provider/, transform/}` move to
`provider/bikeshed/src`.

Stubdata moves to `provider/data/{component}/stubdata`, as it's a
quality-of-life thing and not required for testing or reviewing the
`DatagenProvider` (also, `../stubdata` is a much nicer value for
`ICU4X_DATA_DIR` than `../../../bikeshed/tests/data/baked`). JSON
testdata moves to `provider/bikeshed/data/debug`, as it's used only for
debugging/reviewing, and shouldn't be in `tests/data`, which is data
that actually affects tests.