Update cpp docs to use resolver = 2 (#5122)
Unfortunately this means you can't set features from the CLI anymore,
but I think that's fine.

Without `resolver = 2`, the dependencies from `diplomat` are
feature-unified with the rest, which means that serde gets built with
`std`.


This _may_ be unnecessary with
https://github.com/rust-diplomat/diplomat/pull/510, though I think
setting resolver = 2 here is good practice anyway and we should just do
it.

---------

Co-authored-by: Shane F. Carr <shane@unicode.org>