Reworking filter provider (#5148)
There are a few things here:
* The filters should work on `DataIdentifierBorrowed` instead of
`DataRequest`; metadata should not be relevant for filtering
* This implies a rename from `RequestFilterDataProvider` to
`FilterDataProvider`
* previously the provider implementations returned
`DataErrorKind::Filtered` for filtered ids, however this does not work
with fallback. They should return `DataErrorKind::IdentiferNotFound`; to
a caller (like the fallback adapter) it should be opaque whether the
pipeline contains a filter or not
* `filter_by_langid_allowlist_strict` is marked as "will be replaced
with a smarter algorithm for locale filtering". This smarter algorithm
now exists, so this should be deleted (fixes #834)
* The `Filterable` blanket trait adds `.filterable` to every type in
this crate's rustdoc, and to every type in dev docs. I replaced it with
a normal constructor.