Remove implicit `LocaleFallbackProvider` constructors (#5183)
Now that `LocaleFallbacker` is an explicit argument to `ExportDriver`,
it might make sense to always make it explicit on the
`LocaleFallbackProvider` as well.

The `_unstable`, `_with_buffer_provider`, and `_with_any_provider`
constructors are a bit odd for this crate anyway, no other adapter has
the constructor trinity. Removing them also has the added benefit of
removing the `serde` feature from the crate.