Change pattern representation of fractional seconds; remove obsolete FieldLength variants (#5392)
#1317

This makes it such that adding or removing fractional second digits
involves changing one FieldSymbol, rather than going and mutating the
choice of fields in the pattern. Simpler and less error-prone at
runtime.

The parsing of "ss.SSS" is moved into the parser, where it belongs,
rather than handling it at format time.