Add VarTuple and VarTupleULE (#5511)
This abstraction makes things much easier. It is easier to debug than
`make_varule`, and it composes nicely. I really want to use this for the
plurals and patterns things.

I had to use `transmute_copy` because the compiler doesn't know that the
generic is a fat pointer. But, we know it is, because it is a VarULE
type. I don't think there is anything unsound about this PR.