In zerovec, skip `check_sizes()` test on non-64-bit arches (#5606)
Since the expected sizes are based on 64-bit architectures, we need to
either maintain additional expected sizes for 32-bit architectures, or
(as implemented in this commit) just skip this test on architectures
that are not 64-bit.

Fixes #5605.

<!--
Thank you for your pull request to ICU4X!

Reminder: try to use [Conventional
Comments](https://conventionalcomments.org/) to make comments clearer.

Please see
https://github.com/unicode-org/icu4x/blob/main/CONTRIBUTING.md for
general
information on contributing to ICU4X.
-->