Make time zone parsing non-fallible (#5665)
Based on #5666 

Zones that are added to TZDB in the future should parse as
`Etc/Unknown`. This means every unrecognized string should parse as
`Etc/Unknown`. This then means that the `time_zone_id` field can be
non-optional, as we can always use `Etc/Unknown` as the absent value.

The `metazone_id` field on the other hand need to be double optional:
once for not-resolved, and once for no-metazone-for-zone. This change
improves error cases in datetime formatting, i.e. `Some(None)` is a
UTS-35 fallback, whereas `None` is a `MissingInputField`.