Optimize YearNames (#5721)
Fixes #5602


I didn't try and further squash Japanese with the proposed scheme in
#5602; I'm not really a fan of doing it that way because it seems hacky.
Could be convinced otherwise. Either way, that's a change that can be
done in a followup.

An interesting delta in the fingerprint is that this leads to more
deduplication in Japanese. I'm not sure how.

<!--
Thank you for your pull request to ICU4X!

Reminder: try to use [Conventional
Comments](https://conventionalcomments.org/) to make comments clearer.

Please see
https://github.com/unicode-org/icu4x/blob/main/CONTRIBUTING.md for
general
information on contributing to ICU4X.
-->