Deduplicate tz locations against root (#5759)
A lot of location names are the same in many languages. 

For exemplar cities we have root data, however for territories we don't.
For territories I'm resolving the `und` name by using a Latin-script
endonym (i.e. for region `XY`, resolve `und-Latn-XY`), and if that
fails, just use English (not perfect, but it's better than the raw
region code for both display purposes and deduplication).