Bump versions of semver-broken crates to `{N+1}-dev` (#5769)
I think we should in the future do this as soon as we make a
semver-breaking change anywhere, as it's much harder to do this after
the fact. We can use the `-dev` tag to avoid bumping a crate twice, and
it's easy to remove the tag before a release (or replace with `beta`).