Use `core::error` (#5973)
Fixes #5964 

The only remaining uses of `std` are:
* `DataError::Io` and `DataError::with_path`
* `with_path` can be replaced by `with_display_context(path.display())`,
but the `std::io::ErrorKind` is not in `core` (it could be, it's a
fieldless enum).
* `icu_provider::log` falling back to `println!` (negative `logging`
feature)
* `icu_capi` keeps its std feature, as that's required to build a static
library with panic=unwind. I'm not sure what the use case for this is,
but we use it as a baseline in the size benchmark.