Add POSIX locale parser/converter logic
As per the discussion in #6028, I have implemented POSIX locale parsing
functionality, and a `try_convert_lossy()` function to attempt converting
to a `icu_locale::Locale`. This code is currently in the private `parse`
module, as a temporary solution while support for other platforms is
added and the code structure is finalized.