preserve stucture of timezone designation list (#5581)
The intention of this PR is to maintain the structure of the binary data
in the `DataBlock` struct. The current code produces a list of timezone
designations that has the same length as the list of local time records
with the i-th timezone designation corresponding to the i-th local time
record. In this context the `idx` field of the local time record does
not correspond to a position in the timezone designation list.

In this alternative implementation the timezone designation list
directly corresponds to the null separated list of designations in the
binary data. Since the `idx` field of each local time record can refer
anywhere in that string, even in the middle of a designation, a helper
method is added so that the correct substring is returned.

This allows the `DataBlock` to capture the data exactly and also
potentially re-encode them without loss.

---------

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>