Baked data: use VarULE to store data when specified (#6133)
#5230

I did it by creating a trait, `MaybeAsVarULE`, which is implemented on
data structs, similar to how `Bake` and `Serialize` are implemented on
data structs.

The trait has an associated type that I can leverage in the baked
exporter without having to do `dyn Any` downcasting.

I tried to keep my understanding of @robertbastian's position in mind
while writing this PR (don't add anything hyper-specific to bake in the
provider crate), and I think I achieved this, since the thing I added is
on the same level as the existing format-specific traits.

See the hello world data for an example. I didn't apply it to the other
data markers yet.

---------

Co-authored-by: Robert Bastian <4706271+robertbastian@users.noreply.github.com>