Add TimeZoneParser to `ixdtf` and adjustments to `UtcOffsetRecord` (#6283)
This PR contains a few adjustments to `ixdtf` along with related changes
to `icu_time` for the adjustments

The main changes are:
1. A `TimeZoneParser` was added to `ixdtf` to expose the time zone
parsing logic. Currently some small parsers were implemented in
`temporal_rs`, which leads to differing parsing behavior, so the main
goal is to centralize the parsing logic.
2. Change `UtcOffsetRecord` to an enum, and add `MinutePrecisionOffset`
and `FullPrecisionOffset`. Currently, its technically impossible to tell
the difference between minute precision offsets and full precision
offsets with the current representation. This is primarily meant to
address that issue.
3. A little more clean up of parsing logic on offsets for 1 and 2, plus
some additional unit tests.
4. Adds a `timezone` feature flag to `ixdtf` to gate the
`TimeZoneParser`.