Use regions, not languages, for calendar algorithm defaults (#6325)
Depends on #6302

Matching on languages is incorrect here, these are region codes
(treating `"sa"` as a language code gives us the obviously incorrect
result of mapping Sanskrit to the Islamic calendar). Calendar choice is
corellated with the region, not the language.

This does not start using fallback (though we really should). This also
removes the sa/umalqura mapping entirely since @sffc said it is not in
the latest CLDR.





<!--
Thank you for your pull request to ICU4X!

Reminder: try to use [Conventional
Comments](https://conventionalcomments.org/) to make comments clearer.

Please see
https://github.com/unicode-org/icu4x/blob/main/CONTRIBUTING.md for
general
information on contributing to ICU4X.
-->