Expose Date to/from RataDie (#6369)
Currently we use ISO dates as the "exchange format". For example, when
converting between calendars, we convert through ISO. Most calendars
implement ISO conversion through Rata Die, so we should skip that step.

It's also useful to expose this publicly. Right now we expose to/from
fixed for ISO only, but the code is there to do this for every calendar.