Simplify EraYear (#6443)
This removes the duplication between `standard_era` and
`formatting_era`.

Effectively, whenever `formatting_era` is `FormattingEra::Code`, it is
the same code as `standard_era`, so it could just be `Option<(u8,
TinyStr16)>`. But the fallback era name isn't really needed, as it's
always the same as code up to capitalization, and we don't really care
about capitalization in gigo paths.

The new `era_index` field can be documented without deferring to
whatever datagen does.