Mark Calendar UnstableSealed (#6483)
Note: I used a different header than in datetime for the trait, since I
think unlike datetime this is a non-scaffolding trait that people will
reference often.

Personally my rule of thumb for UnstableSealed is that we expect people
to want to implement it (check) and we have some idea that this is the
direction in which we want to go for custom impls. I hadn't put full
thought into that in the past for this crate but with some of the
formatting stuff split out I'm more confident. So I'm fine with this
being unstable sealed.


I'm not sure if UnstableSealed should live in the cal module or at the
top level.



<!--
Thank you for your pull request to ICU4X!

Reminder: try to use [Conventional
Comments](https://conventionalcomments.org/) to make comments clearer.

Please see
https://github.com/unicode-org/icu4x/blob/main/CONTRIBUTING.md for
general
information on contributing to ICU4X.
-->