Port sort key code from ICU4C
I mostly translate the C++ code straight across, but I made some minor
changes to match the mindset of the Rust codebase, to match Rust
practice, to satisfy Clippy, and to reduce copy-and-paste.  I mostly
avoided slice indexing, except in two places.

A comment on the issue suggested handling the identical level by
appending UTF-8.  That's not what ICU4C does, but it's a whole lot
simpler, so that's what I did.

Fixes #2689